move components into subdirectories
