+++ frontend/src/app/app.module.ts
@@ -5,17 +5,17 @@ import { NgModule } from '@angular/core';
 
 import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing/app-routing.module';
-import { LobbyListItemComponent } from './lobby-list-item/lobby-list-item.component';
+import { LobbyListItemComponent } from './lobby/lobby-list-item/lobby-list-item.component';
 import { SignInComponent } from './sign-in/sign-in.component';
 import { LobbyComponent } from './lobby/lobby.component';
-import { PartyCreateComponent } from './party-create/party-create.component';
+import { PartyCreateComponent } from './lobby//party-create/party-create.component';
 import { PartyComponent } from './party/party.component';
-import { PartyChoosingMenuComponent } from './party-choosing-menu/party-choosing-menu.component';
-import { PartyChoosingRestaurantComponent } from './party-choosing-restaurant/party-choosing-restaurant.component';
-import { PartyOrderingComponent } from './party-ordering/party-ordering.component';
-import { PartyOrderedComponent } from './party-ordered/party-ordered.component';
-import { PartyPaymentComponent } from './party-payment/party-payment.component';
-import { SelectMenuComponent } from './party-choosing-menu/select-menu/select-menu.component';
+import { PartyChoosingMenuComponent } from './party/party-choosing-menu/party-choosing-menu.component';
+import { PartyChoosingRestaurantComponent } from './party/party-choosing-restaurant/party-choosing-restaurant.component';
+import { PartyOrderingComponent } from './party/party-ordering/party-ordering.component';
+import { PartyOrderedComponent } from './party/party-ordered/party-ordered.component';
+import { PartyPaymentComponent } from './party/party-payment/party-payment.component';
+import { SelectMenuComponent } from './party/party-choosing-menu/select-menu/select-menu.component';
 
 @NgModule({
   declarations: [
