+++ frontend/src/app/lobby/lobby-list-item/lobby-list-item.component.html
