+++ frontend/src/app/party/party-choosing-menu/party-choosing-menu.component.html
