+++ frontend/src/app/party/party-choosing-menu/select-menu/select-menu.component.html
