+++ frontend/src/app/party/party-choosing-restaurant/party-choosing-restaurant.component.spec.ts
