+++ frontend/src/app/party/party-ordered/party-ordered.component.ts
